I used this article for this commit guidance: https://www.dummies.com/article/technology/programming-web-design/java/how-to-catch-exceptions-in-java-153217
Here I am focused on a chance for validation of what seems to be user input or input from the DB for questions. I added a try catch block that can be used to validate if the post parameter "questionToAdd" is a valid or invalid question. An exception block could be added if this is something the group intends on pursuing - that would be able to describe why the question may be invalid